# artifacts/api-server/src/routes/bots.ts
import { Router } from "express";
import { v4 as uuidv4 } from "uuid";
import { db, botsTable, botCommandsTable } from "@workspace/db";
import { eq, and } from "drizzle-orm";
import { requireAuth, type AuthRequest } from "../lib/auth.js";

const router = Router();

router.get("/", requireAuth, async (req: AuthRequest, res) => {
  try {
    const bots = await db.select().from(botsTable).where(eq(botsTable.userId, req.userId!));
    res.json(bots.map(formatBot));
  } catch (err) {
    req.log.error({ err }, "List bots error");
    res.status(500).json({ message: "Erro interno" });
  }
});

router.post("/", requireAuth, async (req: AuthRequest, res) => {
  try {
    const { name } = req.body as { name?: string };
    if (!name) {
      res.status(400).json({ message: "Nome do bot é obrigatório" });
      return;
    }
    const [bot] = await db
      .insert(botsTable)
      .values({
        id: uuidv4(),
        userId: req.userId!,
        name,
        status: "disconnected",
        totalGroups: 0,
      })
      .returning();
    res.status(201).json(formatBot(bot));
  } catch (err) {
    req.log.error({ err }, "Create bot error");
    res.status(500).json({ message: "Erro interno" });
  }
});

router.get("/:botId", requireAuth, async (req: AuthRequest, res) => {
  try {
    const { botId } = req.params as { botId: string };
    const [bot] = await db
      .select()
      .from(botsTable)
      .where(and(eq(botsTable.id, botId), eq(botsTable.userId, req.userId!)));
    if (!bot) {
      res.status(404).json({ message: "Bot não encontrado" });
      return;
    }
    res.json(formatBot(bot));
  } catch (err) {
    req.log.error({ err }, "Get bot error");
    res.status(500).json({ message: "Erro interno" });
  }
});

router.delete("/:botId", requireAuth, async (req: AuthRequest, res) => {
  try {
    const { botId } = req.params as { botId: string };
    const [bot] = await db
      .select()
      .from(botsTable)
      .where(and(eq(botsTable.id, botId), eq(botsTable.userId, req.userId!)));
    if (!bot) {
      res.status(404).json({ message: "Bot não encontrado" });
      return;
    }
    await db.delete(botsTable).where(eq(botsTable.id, botId));
    res.status(204).send();
  } catch (err) {
    req.log.error({ err }, "Delete bot error");
    res.status(500).json({ message: "Erro interno" });
  }
});

router.post("/:botId/connect", requireAuth, async (req: AuthRequest, res) => {
  try {
    const { botId } = req.params as { botId: string };
    const { type, phone } = req.body as { type?: string; phone?: string };

    const [bot] = await db
      .select()
      .from(botsTable)
      .where(and(eq(botsTable.id, botId), eq(botsTable.userId, req.userId!)));
    if (!bot) {
      res.status(404).json({ message: "Bot não encontrado" });
      return;
    }

    const sessionId = uuidv4();

    if (type === "qrcode") {
      const mockQrData = `2@${uuidv4()},${Date.now()},botaluguel-pro-connect-${botId}`;
      await db
        .update(botsTable)
        .set({ status: "connecting", connectionType: "qrcode", qrCode: mockQrData, pairCode: null, phone: phone || null })
        .where(eq(botsTable.id, botId));

      res.json({ type: "qrcode", qrCode: mockQrData, pairCode: null, sessionId });
    } else if (type === "code") {
      const pairCode = Math.random().toString(36).substring(2, 10).toUpperCase();
      await db
        .update(botsTable)
        .set({ status: "connecting", connectionType: "code", pairCode, qrCode: null, phone: phone || null })
        .where(eq(botsTable.id, botId));

      res.json({ type: "code", qrCode: null, pairCode, sessionId });
    } else {
      res.status(400).json({ message: "Tipo de conexão inválido. Use 'qrcode' ou 'code'" });
    }
  } catch (err) {
    req.log.error({ err }, "Connect bot error");
    res.status(500).json({ message: "Erro interno" });
  }
});

router.post("/:botId/disconnect", requireAuth, async (req: AuthRequest, res) => {
  try {
    const { botId } = req.params as { botId: string };
    const [bot] = await db
      .select()
      .from(botsTable)
      .where(and(eq(botsTable.id, botId), eq(botsTable.userId, req.userId!)));
    if (!bot) {
      res.status(404).json({ message: "Bot não encontrado" });
      return;
    }
    const [updated] = await db
      .update(botsTable)
      .set({ status: "disconnected", qrCode: null, pairCode: null, connectedAt: null })
      .where(eq(botsTable.id, botId))
      .returning();
    res.json(formatBot(updated));
  } catch (err) {
    req.log.error({ err }, "Disconnect bot error");
    res.status(500).json({ message: "Erro interno" });
  }
});

router.get("/:botId/commands", requireAuth, async (req: AuthRequest, res) => {
  try {
    const { botId } = req.params as { botId: string };
    const [bot] = await db
      .select()
      .from(botsTable)
      .where(and(eq(botsTable.id, botId), eq(botsTable.userId, req.userId!)));
    if (!bot) {
      res.status(404).json({ message: "Bot não encontrado" });
      return;
    }
    let [commands] = await db.select().from(botCommandsTable).where(eq(botCommandsTable.botId, botId));
    if (!commands) {
      const [created] = await db
        .insert(botCommandsTable)
        .values({ id: uuidv4(), botId, nodes: [], edges: [] })
        .returning();
      commands = created;
    }
    res.json({
      botId: commands.botId,
      nodes: commands.nodes,
      edges: commands.edges,
      updatedAt: commands.updatedAt,
    });
  } catch (err) {
    req.log.error({ err }, "Get bot commands error");
    res.status(500).json({ message: "Erro interno" });
  }
});

router.put("/:botId/commands", requireAuth, async (req: AuthRequest, res) => {
  try {
    const { botId } = req.params as { botId: string };
    const { nodes, edges } = req.body as { nodes?: unknown[]; edges?: unknown[] };
    const [bot] = await db
      .select()
      .from(botsTable)
      .where(and(eq(botsTable.id, botId), eq(botsTable.userId, req.userId!)));
    if (!bot) {
      res.status(404).json({ message: "Bot não encontrado" });
      return;
    }
    const existing = await db.select().from(botCommandsTable).where(eq(botCommandsTable.botId, botId));
    let result;
    if (existing.length === 0) {
      [result] = await db
        .insert(botCommandsTable)
        .values({ id: uuidv4(), botId, nodes: nodes ?? [], edges: edges ?? [] })
        .returning();
    } else {
      [result] = await db
        .update(botCommandsTable)
        .set({ nodes: nodes ?? [], edges: edges ?? [], updatedAt: new Date() })
        .where(eq(botCommandsTable.botId, botId))
        .returning();
    }
    res.json({
      botId: result.botId,
      nodes: result.nodes,
      edges: result.edges,
      updatedAt: result.updatedAt,
    });
  } catch (err) {
    req.log.error({ err }, "Save bot commands error");
    res.status(500).json({ message: "Erro interno" });
  }
});

function formatBot(bot: typeof botsTable.$inferSelect) {
  return {
    id: bot.id,
    userId: bot.userId,
    name: bot.name,
    phone: bot.phone,
    status: bot.status,
    connectionType: bot.connectionType,
    qrCode: bot.qrCode,
    pairCode: bot.pairCode,
    totalGroups: bot.totalGroups,
    createdAt: bot.createdAt,
    connectedAt: bot.connectedAt,
  };
}

export default router;